backend/requirements.txt
flask==2.3.3
mysql-connector-python==8.4.0
oss2==2.18.6
python-multipart==0.0.9